C API の `mp4_file_muxer_set_reserved_moov_box_size()` の `size` 引数の型を `u64` から `u32` に変更する

+++ crates/c-api/include/mp4.h
@@ -1454,7 +1454,7 @@ const char *mp4_file_muxer_get_last_error(const struct Mp4FileMuxer *muxer);
  * ```
  */
 enum Mp4Error mp4_file_muxer_set_reserved_moov_box_size(struct Mp4FileMuxer *muxer,
-                                                        uint64_t size);
+                                                        uint32_t size);
 
 /**
  * MP4 ファイルのマルチプレックス処理を初期化する
